sketch incorporation of github CLI support (#198)

+++ cargo-smart-release/src/changelog/section/segment.rs
@@ -130,7 +130,7 @@ impl ThanksClippy {
 }
 
 bitflags! {
-    pub struct Selection: u32 {
+    pub struct Selection: u8 {
         const CLIPPY = 1<<0;
         const COMMIT_DETAILS = 1<<1;
         const COMMIT_STATISTICS = 1<<2;
